+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * CONRAT-42855: Update deprecated URLSession initializer for unit test mock objects
 * CONRAT-43326: Re-enable SwiftLint.
 * CONRAT-43324: Remove NSObject.ratTracker from AnalyticsManager and use RAnalyticsRATTracker shared instance instead.
+* CONRAT-43283: Remove AnalyticsManager dependency from RAnalyticsDatabase.
 
 ### Build fixes
 * CONRAT-43264: fix the Swift Package Test Target.